src/requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
blis==0.7.4
bravado==11.0.3
bravado-core==5.17.0
brotlipy==0.7.0
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
cheroot==8.5.2
CherryPy==18.6.0
click==7.1.2
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
dnspython==2.1.0
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fire==0.4.0
fsspec==2021.7.0
future==0.18.2
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.18
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
humanize==3.2.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
ipython @ file:///tmp/build/80754af9/ipython_1610724799192/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2 @ file:///home/linux1/recipes/ci/jinja2_1610990516718/work
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
monotonic==1.6
more-itertools==8.7.0
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
neptune-client==0.10.2
neptune-pytorch-lightning==0.9.7
networkx==2.5
nltk==3.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.0
olefile==0.46
onnx==1.8.1
packaging==20.9
pandas==1.2.1
parso==0.7.0
pathy==0.4.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pkginfo==1.7.0
portend==2.7.1
preshed==3.0.5
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.15.2
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.3
pyDeprecate==0.3.0
pydicom==2.1.2
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyJWT==2.1.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.1
python-etcd==0.4.5
pytorch-lightning==1.3.8
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.11.13
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
simplejson==3.17.3
six @ file:///tmp/build/80754af9/six_1605205327372/work
smart-open==3.0.0
smmap==4.0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
spacy==3.0.1
spacy-legacy==3.0.1
srsly==2.4.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tempora==4.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thinc==8.0.1
threadpoolctl==2.1.0
tifffile==2021.2.1
torch==1.7.1
torchelastic==0.2.1
torchmetrics==0.4.1
torchvision==0.8.2
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typer==0.3.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wasabi==0.8.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webcolors==1.11.1
websocket-client==1.1.0
Werkzeug==1.0.1
yarl==1.6.3
zc.lockfile==2.0
